new attack on single ssh-agent use
we generate a key value based on user, epoch, machine and use that to
link to the loadfile.  if the link is good we load the file.  then if
ssh-add shows keys we are done if not then we add the keys.  when we log
out if the keyfile is a link we remove it.  if the loadfile is not a link
then we run ssh-agent -k and remove the loadfile.