more tweaks.
consolidated the USER0? to _LOAD01 file.  adjusted bloader to use a
perl script to deal with comments and lines starting with non-word
chars. this make inline documentation easier and turning functions on
and off as easy as a non-word (incl space) first char on a line.
bashrc updated for _LOAD01.  tried to remove uselss heading and trailing
blank lines in files.  makes md5sum check/comparison possible.
major update to Makefile to be streamlined and pretty.