this is a bulk update
most of this is from work ideas. how to make sure you are using the
latest installed bash. system can be out of data wrt an installed
version. so the idea is to grab the list of installed bashes, append
a version number, sort numerically and take the first value. then we
grab the version of the current bash ($SHELL) and if they are the same
we are done. otherwise we exec the more recent version. along with this
update were changes to final, bashrcfuncs (somethings were always ro
which needed to change). finally we made some updates to vim so the
status line was easier to read and present.