updates to Prompt
Lots of changes. this adds a lot of flexibility. now you can use a
command to turn on/off settings, set text to values and change colors.
this updates cprompt, removes aliases tag and xtra, adds functions tag
and extra, and is a complete rewrite of _SetPrompt. we also added some
maps for prompt control in prompt_conf and simplified the assignment of
the host color. right now all of this is changable by user but we force
host and prompt to always be shown. we can do the same thing to force a
fixed color for prompt and host for the root user if needed. just set
a readonly var to the content we want to keep then reassign that value
to the map values respectively.