made some name changes to rules
it seems clearer to use 'make update' to generate the updated repo. and
as a general use thing, i've started using a stub in the ~/Env dir to
include the one in the git repo. that way i can just source that one and
if it changes i get the most current version.