shuffle things around some. skeletons and closets
most of this is housecleaning. since we are setting rx for functions
in the bashrc we remove that declare from bashrcfuncs. the make file
now allows us to install things (make install-bashrc) and the install
items are in dotinstall (moved dotcopies to dotinstall). the install
bashrc is simpler. we now have timeouts for login (testing times out
after 3 sec)