work with ssh-agent setup
i'll try this. turns out that a regular xterm was not processing the
.bash_logout file. missed the '-ls' option. fixed that. so now the
.bash_logout is getting read. so updated that (installed fortune)
and am updating the sshAgentPrep script so that if the file in /tmp is
missing the users .agentfile (?) is removed and the var is unset. then
it gets generated and saved. after that the tmp file is sourced and
used. hoping to get the sshagent stuff to persist across multiple
xterms. that way i only need to enter the keys once and things should
persist until the /tmp ref goes away. at worst we just remove the users
file by hand and rerun the sshAgentPrep script.