more improvements to Makefile and bashrc
time to prune stuff. removed the priv from dotinstall so i needed to
remove the references in bashrc to that. this leaves FDIRLIST as a
single value. i may further modify that to just the fullpath for FLIST.
the Makefile has been getting a lot fo work. we now can do installs of
all the dotfiles, individual dotfiles, and can check the dotfiles. some
minor rearrangement of the targets to help structure. the default target
is now help.