cleanup and fixing hlp
so we removed the stp alias (setPrompt -> stp) and cleaned that up.
then we worked on hlp so that with no args it provides the first help
line for each function and full help if given a function name. then we
used csplit to break this up into parts, renamed the parts to get
alphabetical order and cat'd them to produce the new functions file.
getting closer to real testing.