found out something about bash....
if you want to remove a fixed prefix string, just do x=prefix then
echo ${var#$x/} and you get everything after the prefix. neat. putting
that into cksumit to get relative paths for files. updated Makefile to
figure out the prune value and that gets passed in to cksumit on the
command line. if not specified this should just remove the leading /.