updates to fsplit and fjoin
splitting on blank lines was causing problems so this should fix it.
we use #split# as a token to split on, removing that token after the
files are split and removing empty files (stat -c %s). maybe better to
just use 'find ... -type empty' and remove all in one fell swoop.
then we modify fjoin to cat the files with inserting '#split#' after
each file cat'd. i'm being clever with <(echo '#split#') as an arg to
cat but it might be faster to just cat; echo; on two lines. will make
these changes next.