work on a better stp func
so we wanted something more flexible. too much to remember with the old
way and too complex to code around. so we are using key arguments to
define input values. this affects stp (rewrite) and base_init (initial
variable definitions). we also rename the variables (blah_ to _blah) and
update the extra.rc and prime.rc files. maybe need to do this with
tools.rc but cross that bridge later.