base_init: combined some inits into one line, fixed a typo
prime.rc: improved logic in _ConverColor, typo in _ShowPrompt,
error in accessing index keys for _InitialText, overall logic in stp's
cascaded if's, support using current values as defaults unless changed on
command line, removed confusion over keys (indexes of _InitialText) and the
renamed keys (making keys a pat instead of a list). we did this by declaring
keypat to be the old subbed value of keys.